Extract types and interfaces from <TransactionView /> to TransactionFilterTypes.ts
